Fix BookRoom_04, Add BookRoom_05

+++ HotelTelegramBot/Controller/MessageController.cs
@@ -45,14 +45,26 @@ namespace HotelTelegramBot.Controller
         {
             Chat chat = e.CallbackQuery.Message.Chat;
             int messageId = e.CallbackQuery.Message.MessageId;
-            string chatPosition = DbServices.GetChatPositionByIdChat(chat.Id);
-            string userInput = e.CallbackQuery.Data;
 
-            Logger.Log(chatPosition, e);
+            if (chatResponder == null)
+            {
+                chatResponder = new ChatResponder(new Start(e.CallbackQuery.Message.Chat));
+                return;
+            }
 
             await Program.botClient.DeleteMessageAsync(chat, messageId);
-            await ServicesMessageController.RouteMenuAsync(userInput, chat);
-            await ServicesMessageController.RouteMessageChatPositionAsync(chatPosition, e);
+            chatResponder.ReceiveMessageAsync(e);
+
+            //Chat chat = e.CallbackQuery.Message.Chat;
+            //int messageId = e.CallbackQuery.Message.MessageId;
+            //string chatPosition = DbServices.GetChatPositionByIdChat(chat.Id);
+            //string userInput = e.CallbackQuery.Data;
+
+            //Logger.Log(chatPosition, e);
+
+            //await Program.botClient.DeleteMessageAsync(chat, messageId);
+            //await ServicesMessageController.RouteMenuAsync(userInput, chat);
+            //await ServicesMessageController.RouteMessageChatPositionAsync(chatPosition, e);
         }
     }
 }
